memory.hpp replaced in dir include
